Update to Sections
I got the sections page to display: crn, course_name, days, start_time, end_time, status, enroll_avail, and enroll_tot, made those sections selectable, and made it where you could click that button to add them to the cart. I think it would look better in a pho-table format, but I will mess with that tomorrow. I am not sure how to tell the the button which page it is going to, but it should be able to import that info into the cart you are making Jon.